+++ News/Controllers/UsersController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using News.Entities.Models.ModelsUpdate;
 using News.Services.ServicesInterface;
+using System.Xml.XPath;
 
 namespace News.Controllers;
 
@@ -22,6 +23,10 @@ public class UsersController : ControllerBase
     public async Task<IActionResult> GetAsync([FromRoute] Guid id)
     {
         var user = await _service.GetAsync(id);
+        if (user is null)
+        {
+            return NotFound($"User with ID {id} was not found.");
+        }
 
         return Ok(user);
     }
@@ -29,6 +34,17 @@ public class UsersController : ControllerBase
     [HttpPatch("{id}")]
     public async Task<IActionResult> UpdateAsync([FromRoute] string id, [FromBody] JsonPatchDocument<UserUpdateModel> user)
     {
+        if (user is null)
+        {
+            return BadRequest("Patch document cannot be null.");
+        }
+
+        var existingUser = await _service.GetAsync(Guid.Parse(id));
+        if (existingUser is null)
+        {
+            return NotFound($"User with ID {id} was not found.");
+        }
+
         var updatedUser = await _service.UpdateAsync(user, id);
 
         return Ok(updatedUser);
@@ -37,6 +53,12 @@ public class UsersController : ControllerBase
     [HttpDelete]
     public async Task<IActionResult> DeleteAsync([FromQuery] Guid id)
     {
+        var existingUser = await _service.GetAsync(id);
+        if (existingUser is null)
+        {
+            return NotFound($"User with ID {id} was not found.");
+        }
+
         await _service.DeleteAsync(id);
 
         return Ok();
